calibration/requirements.txt
opencv-python
numpy
PyYAML
glob2